code-examples/chapter-05/progressive/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pyjwt>=2.8.0

